requirements.txt
customtkinter>=5.2.0
cryptography>=42.0.0
pyperclip>=1.8.2
Pillow>=10.0.0